유동민 : 오류 수정

+++ 3DGameProject/GameEngineContents/BaseEnemyActor.h
@@ -140,7 +140,7 @@ protected:
 	std::function<void()> SuperArmorOn_Callback = nullptr;
 	std::function<void()> SuperArmorOff_Callback = nullptr;
 
-	int RotateValue = 0;
+	float RotateValue = 0.0f;
 	void ChasePlayer(float _DeltaTime);
 	void SuperArmorOn();
 	void SuperArmorOff();
